+++ tests/PluginTest.php
@@ -3,14 +3,17 @@ declare( strict_types = 1 );
 // phpcs:disable MediaWiki.NamingConventions.ValidGlobalName
 // phpcs:disable MediaWiki.Commenting.MissingCovers.MissingCovers -- T363064
 
+use MediaWikiPhanConfig\Plugin\FirstClassCallableRecommendFixer;
 use Phan\CLIBuilder;
 use Phan\CodeBase;
 use Phan\Config;
 use Phan\Language\Scope\GlobalScope;
 use Phan\Language\Type;
+use Phan\Library\FileCache;
 use Phan\Output\Printer\PlainTextPrinter;
 use Phan\Phan;
 use Phan\Plugin\ConfigPluginSet;
+use Phan\Plugin\Internal\IssueFixingPlugin\IssueFixer;
 use PHPUnit\Framework\Attributes\CoversNothing;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Output\BufferedOutput;
@@ -60,9 +63,9 @@ class PluginTest extends TestCase {
 	 * @param string $plugin
 	 * @param string $cfgFile
 	 * @param bool $usePolyfill Whether to force the polyfill parser
-	 * @return string|null
+	 * @return array{?string, ?string}
 	 */
-	private function runPhan( string $plugin, string $cfgFile, bool $usePolyfill ): string {
+	private function runPhan( string $plugin, string $cfgFile, bool $usePolyfill ): array {
 		if ( !$usePolyfill && !extension_loaded( 'ast' ) ) {
 			$this->markTestSkipped( 'This test requires PHP extension \'ast\' loaded' );
 		}
@@ -73,6 +76,7 @@ class PluginTest extends TestCase {
 		$cliBuilder->setOption( 'project-root-directory', __DIR__ );
 		$cliBuilder->setOption( 'config-file', $cfgFile );
 		$cliBuilder->setOption( 'directory', "./plugins/$plugin" );
+		$cliBuilder->setOption( 'exclude-file', "plugins/$plugin/fixed.php" );
 		$cliBuilder->setOption( 'no-progress-bar', true );
 		if ( $usePolyfill ) {
 			$cliBuilder->setOption( 'force-polyfill-parser', true );
@@ -89,11 +93,68 @@ class PluginTest extends TestCase {
 		$printer->configureOutput( $stream );
 		Phan::setPrinter( $printer );
 
+		// We need to intercept the issue collector like this, because they somehow disappear
+		// if you just call Phan::getIssueCollector()->getCollectedIssues() at the end.
+		// There has to be a better way to do this...
+		$instances = [];
+		Phan::setIssueCollector( new class( $instances ) implements \Phan\Output\IssueCollectorInterface {
+			private $instances;
+			private $realCollector;
+
+			public function __construct( &$instances ) {
+				$this->instances =& $instances;
+				$this->realCollector = new \Phan\Output\Collector\BufferingCollector;
+			}
+
+			public function collectIssue( ...$arguments ): void {
+				$this->realCollector->collectIssue( ...$arguments );
+			}
+
+			public function getCollectedIssues( ...$arguments ): array {
+				$this->instances = $this->realCollector->getCollectedIssues();
+				return $this->realCollector->getCollectedIssues( ...$arguments );
+			}
+
+			public function removeIssuesForFiles( ...$arguments ): void {
+				$this->realCollector->removeIssuesForFiles( ...$arguments );
+			}
+
+			public function reset( ...$arguments ): void {
+				$this->instances = $this->realCollector->getCollectedIssues();
+				$this->realCollector->reset( ...$arguments );
+			}
+
+			public function flush(): void {
+				$this->instances = $this->realCollector->getCollectedIssues();
+				$this->realCollector->flush();
+			}
+		} );
+
 		Phan::analyzeFileList( $this->codeBase, static function () use ( $cli ) {
 			return $cli->getFileList();
 		} );
 
-		return $stream->fetch();
+		$issues = $stream->fetch();
+		$fixed = $this->computeFixes( $this->codeBase, $instances );
+
+		return [ $issues, $fixed ];
+	}
+
+	private function computeFixes( $code_base, $instances ): ?string {
+		// TODO This should not be hardcoded, but we only have one fixer now, so whatever
+		IssueFixer::registerFixerClosure( 'MediaWikiUseFirstClassCallable',
+			FirstClassCallableRecommendFixer::fix( ... ), );
+
+		$fixers_for_files = IssueFixer::computeFixersForInstances( $instances );
+
+		foreach ( $fixers_for_files as $file => $fixers ) {
+			$entry = FileCache::getOrReadEntry( $file );
+			$contents = $entry->getContents();
+			$new_contents = IssueFixer::computeNewContentForFixers( $code_base, $file, $contents, $fixers );
+			// Assume only one file exists
+			return $new_contents;
+		}
+		return null;
 	}
 
 	/**
@@ -110,11 +171,13 @@ class PluginTest extends TestCase {
 			$folder = $dir->getPathname();
 			$testName = basename( $folder );
 			$expected = file_get_contents( $folder . '/expectedResults.txt' );
+			$fixed = file_exists( $folder . '/fixed.php' ) ? file_get_contents( $folder . '/fixed.php' ) : null;
 
 			yield "$plugin/$testName" => [
 				"$plugin/$testName",
 				"./plugins/$plugin/config.php",
-				$expected
+				$expected,
+				$fixed
 			];
 		}
 	}
@@ -140,16 +203,20 @@ class PluginTest extends TestCase {
 	 * @param string $path
 	 * @param string $configFile
 	 * @param string $expected
+	 * @param ?string $fixed
 	 */
-	public function testPlugins( string $path, string $configFile, string $expected ): void {
+	public function testPlugins( string $path, string $configFile, string $expected, ?string $fixed ): void {
+		[ $actualIssues, $actualFixed ] = $this->runPhan( $path, $configFile, false );
+
 		// Replace backslashes with slashes and replace CRLF with LF, both appear when running on Windows.
-		$actual = str_replace(
+		$actualIssues = str_replace(
 			[ "\r", '\\' ],
 			[ '', '/' ],
-			$this->runPhan( $path, $configFile, false )
+			$actualIssues
 		);
 
-		static::assertSame( $expected, $actual );
+		static::assertSame( $expected, $actualIssues );
+		static::assertSame( $fixed, $actualFixed );
 	}
 
 	/**
@@ -160,13 +227,17 @@ class PluginTest extends TestCase {
 	 * @param string $expected
 	 */
 	public function testPlugins_Polyfill( string $path, string $configFile, string $expected ): void {
+		if ( $path === 'FirstClassCallableRecommendPlugin/basic' ) {
+			$this->markTestSkipped( "Named params don't work in polyfill parser?" );
+		}
+
 		// Replace backslashes with slashes and replace CRLF with LF, both appear when running on Windows.
-		$actual = str_replace(
+		[ $actualIssues, $actualFixed ] = str_replace(
 			[ "\r", '\\' ],
 			[ '', '/' ],
 			$this->runPhan( $path, $configFile, true )
 		);
 
-		static::assertSame( $expected, $actual );
+		static::assertSame( $expected, $actualIssues );
 	}
 }